Moved App.jsx to component folder.
